First Iteration of notifying lecturer via email
Front-end:
Created function for 'Apply for slot' button (notify lecturer). This includes an alert to notify the user this has happened.

Added function in lab-swap and account service, which posts to the back-end '/notifyLecturer' route on the lab swap and account controller. (Used both for now as Im unsure which service is better to use, for data retrieval reasons)

Removed some unused functions from register component.

Back-end:
Created functions on both lab-swap and account services that contain the email message and instructions to get data and call X function etc.

Created the required routes and functions on both lab-swap and account controllers.