Improve the "Create Lab" Modal layout.
Added curves to modal edges.
Corrected the size of the modal, to better suit the UI.
Introduced bottom margin property, to Modal Header, to put space between the line and text underneath.
Adjusted the column size of each form field.
Made it so the modal opens in the middle of the screen, when activated.